requirements.txt
requests
beautifulsoup4
pytest-asyncio